streamlit_app/requirements.txt
streamlit>=1.34,<2.0
python-dotenv>=1.0,<2.0